Added view model for creating a DayEntry Updated DayEntry entity to reflect changes in DB

+++ Eight.Domian/Models/DayEntry.cs
@@ -8,7 +8,7 @@ namespace Eight.Domian.Models
         public string Description { get; set; }
         public DateTime DateLogged { get; set; }
         public decimal DayRating { get; set; }
-        public byte ActivitiesCompleted { get; set; }
+        public int ActivitiesCompleted { get; set; }
         public int CategoryId { get; set; }
         
         public virtual Category Category { get; set; }
